+++ packages/mobile/src/buttons/SlideButton.tsx
@@ -17,7 +17,16 @@ export const slideButtonTestID = 'slide-button';
 
 export type SlideButtonBackgroundProps = Pick<
   SlideButtonBaseProps,
-  'borderRadius' | 'checked' | 'compact' | 'disabled' | 'uncheckedLabel' | 'variant'
+  | 'borderBottomLeftRadius'
+  | 'borderBottomRightRadius'
+  | 'borderRadius'
+  | 'borderTopLeftRadius'
+  | 'borderTopRightRadius'
+  | 'checked'
+  | 'compact'
+  | 'disabled'
+  | 'uncheckedLabel'
+  | 'variant'
 > & {
   progress: SpringValue<number>;
   style?: StyleProp<ViewStyle>;
@@ -151,6 +160,10 @@ export const SlideButton = memo(
         checked,
         compact,
         borderRadius = compact ? 700 : 900,
+        borderTopLeftRadius,
+        borderTopRightRadius,
+        borderBottomLeftRadius,
+        borderBottomRightRadius,
         uncheckedLabel,
         checkedLabel,
         onSlideStart,
@@ -284,7 +297,11 @@ export const SlideButton = memo(
       return (
         <View ref={ref} id={labelId} onLayout={onLayout} style={containerStyle} testID={testID}>
           <SlideButtonBackgroundComponent
+            borderBottomLeftRadius={borderBottomLeftRadius}
+            borderBottomRightRadius={borderBottomRightRadius}
             borderRadius={borderRadius}
+            borderTopLeftRadius={borderTopLeftRadius}
+            borderTopRightRadius={borderTopRightRadius}
             checked={checked}
             compact={compact}
             disabled={disabled}
@@ -301,7 +318,11 @@ export const SlideButton = memo(
                 {...(typeof accessibilityLabel === 'string'
                   ? { accessibilityLabel }
                   : { accessibilityLabelledBy: labelId })}
+                borderBottomLeftRadius={borderBottomLeftRadius}
+                borderBottomRightRadius={borderBottomRightRadius}
                 borderRadius={borderRadius}
+                borderTopLeftRadius={borderTopLeftRadius}
+                borderTopRightRadius={borderTopRightRadius}
                 checked={checked}
                 checkedLabel={checkedLabel}
                 compact={compact}