+++ packages/web/src/tabs/SegmentedTab.tsx
@@ -2,7 +2,6 @@ import React, { forwardRef, memo, useCallback, useMemo } from 'react';
 import type { ThemeVars } from '@coinbase/cds-common/core/theme';
 import { useTabsContext } from '@coinbase/cds-common/tabs/TabsContext';
 import { type TabValue } from '@coinbase/cds-common/tabs/useTabs';
-import type { SharedProps } from '@coinbase/cds-common/types/SharedProps';
 import { css } from '@linaria/core';
 import { m as motion } from 'framer-motion';
 
@@ -13,7 +12,7 @@ import { Text } from '../typography/Text';
 
 import { tabsTransitionConfig } from './Tabs';
 
-const MotionText = motion(Text);
+const MotionBox = motion(Box);
 
 const insetFocusRingCss = css`
   &:focus {
@@ -136,9 +135,10 @@ const SegmentedTabComponent = memo(
           type="button"
           {...props}
         >
-          <Box as="span" justifyContent="center" paddingX={2} paddingY={1}>
+          <MotionBox as="span" justifyContent="center" paddingX={2} paddingY={1} {...motionProps}>
             {typeof label === 'string' ? (
-              <MotionText
+              <Text
+                color="currentColor"
                 font={font}
                 fontFamily={fontFamily}
                 fontSize={fontSize}
@@ -146,14 +146,13 @@ const SegmentedTabComponent = memo(
                 lineHeight={lineHeight}
                 textAlign={textAlign}
                 textTransform={textTransform}
-                {...motionProps}
               >
                 {label}
-              </MotionText>
+              </Text>
             ) : (
               label
             )}
-          </Box>
+          </MotionBox>
         </Pressable>
       );
     },