+++ packages/mobile-visualization/CHANGELOG.md
@@ -8,7 +8,11 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
-## Unreleased
+## 3.4.0-beta.20 (2/27/2026 PST)
+
+#### 🚀 Updates
+
+- Add styles props to PeriodSelector. [[#438](https://github.com/coinbase/cds/pull/438/)]
 
 #### 📘 Misc